Services created for student

+++ src/main/java/Student/Management/demo/Repository/PersonRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PersonRepo extends JpaRepository<PersonModel,Integer> {
+public interface PersonRepo extends JpaRepository<PersonModel, Long> {
 }
